infra: move types to internal
